Complete integration with Azure Face and Language services

+++ Bff.Core/OutputPorts/EmotionsResult.cs
@@ -11,38 +11,38 @@ namespace Bff.Core.OutputPorts
         private readonly List<string> _negatives = new List<string> { nameof(AngerScore), nameof(DisgustScore), nameof(FearScore), nameof(SadnessScore) };
         private readonly List<string> _positives = new List<string> { nameof(HappinessScore), nameof(SurpriseScore) };
 
-        public decimal AngerScore { get; set; }
+        public double AngerScore { get; set; }
 
-        public decimal ContemptScore { get; set; }
+        public double ContemptScore { get; set; }
 
-        public decimal DisgustScore { get; set; }
+        public double DisgustScore { get; set; }
 
-        public decimal FearScore { get; set; }
+        public double FearScore { get; set; }
 
-        public decimal HappinessScore { get; set; }
+        public double HappinessScore { get; set; }
 
-        public decimal NeutralScore { get; set; }
+        public double NeutralScore { get; set; }
 
-        public decimal SadnessScore { get; set; }
+        public double SadnessScore { get; set; }
 
-        public decimal SurpriseScore { get; set; }
+        public double SurpriseScore { get; set; }
 
         public bool IsNegative => _negatives.Contains(GetMaxScore().Name); 
 
         public bool IsPositive => _positives.Contains(GetMaxScore().Name);
 
-        public (string Name, decimal Value) GetMaxScore()
+        public (string Name, double Value) GetMaxScore()
         {
             var pairs = GetType().GetProperties()
-                                 .Where(p => p.PropertyType == typeof(decimal))
+                                 .Where(p => p.PropertyType == typeof(double))
                                  .Select(p => new
                                  {
                                      Name = p.Name,
-                                     Value = (decimal)p.GetValue(this)
+                                     Value = (double)p.GetValue(this)
                                  });
 
             string maxName = string.Empty;
-            decimal maxValue = 0;
+            double maxValue = 0;
 
             foreach (var item in pairs)
             {
